Introduced abstraction layer in NRTReplicationEngine for extension
